Manual

The PSRDatabase module provides interfaces to access data structured by PSR to be used in its models. Currently there are two main interfaces.

The interface for studies. This interface is designed to read parameters from the files, some examples are deficit costs, fuel costs, currency, storage capacity etc.

The interface for reading and writing time series data. Time series data in the context of most studies have 4 dimensions (agents, stages, scenarios and blocks). Since studies of renewables with multiple agents, scenarios and stages can get quite big, we have designed different formats that are optimized to some objective (human readability, size, fast reading and writing, etc.).

Both interfaces are defined as a set of methods that need to be implemented to make a different file format work. In this manual we will describe the abstract methods and give concrete examples of code to perform the work needed.

When using the PSRDatabase package in your codebase we strongly advise you to create a constant PSRI to keep the code concise and explicitly declare that a certain function came from PSRDatabase. This can be done by adding the following code to the top of the code

using PSRDatabase
const PSRI = PSRDatabase

Initialize Study

Missing docstring.

Missing docstring for PSRDatabase.AbstractData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.AbstractStudyInterface. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.load_study. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.description. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.max_elements. Check Documenter's build log for details.

Study dimensions

Missing docstring.

Missing docstring for PSRDatabase.StageType. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.total_stages. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.total_scenarios. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.total_blocks. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.total_openings. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.total_stages_per_year. Check Documenter's build log for details.

Study duration and blocking

PSRDatabase.BlockDurationMode
PSRDatabase.stage_duration
PSRDatabase.block_duration
PSRDatabase.block_from_stage_hour

Read Scalar Attributes

Missing docstring.

Missing docstring for PSRDatabase.configuration_parameter. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_code. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_name. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_parm. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_parm_1d. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_parms. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_parms_1d. Check Documenter's build log for details.

Read Vector Attributes

Time controller

Missing docstring.

Missing docstring for PSRDatabase.mapped_vector. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.go_to_stage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.go_to_dimension. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.update_vectors!. Check Documenter's build log for details.

Direct access

Missing docstring.

Missing docstring for PSRDatabase.get_vector. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_vector_1d. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_vector_2d. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_vectors. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_vectors_1d. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_vectors_2d. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_nonempty_vector. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_series. Check Documenter's build log for details.

Relations between collections

Missing docstring.

Missing docstring for PSRDatabase.RelationType. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.is_vector_relation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_references. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_vector_references. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_map. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_vector_map. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_reverse_map. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_reverse_vector_map. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_related. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_vector_related. Check Documenter's build log for details.

Reflection

Missing docstring.

Missing docstring for PSRDatabase.get_attribute_dim1. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_attribute_dim2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_collections. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_attributes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_attribute_struct. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_data_struct. Check Documenter's build log for details.

PSRDatabase.Attribute — Type

Attribute

Abstract type for attributes, the building blocks of collections.

Missing docstring.

Missing docstring for PSRDatabase.get_attributes_indexed_by. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_relations. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.get_attribute_dim. Check Documenter's build log for details.

Read and Write Graf files

Open and Close

Missing docstring.

Missing docstring for PSRDatabase.AbstractReader. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.AbstractWriter. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.open. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.close. Check Documenter's build log for details.

Write entire file

Missing docstring.

Missing docstring for PSRDatabase.array_to_file. Check Documenter's build log for details.

Write registry

Missing docstring.

Missing docstring for PSRDatabase.write_registry. Check Documenter's build log for details.

Header information

Missing docstring.

Missing docstring for PSRDatabase.is_hourly. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.hour_discretization. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.max_stages. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.max_scenarios. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.max_blocks. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.max_blocks_current. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.max_blocks_stage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.max_agents. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.stage_type. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.initial_stage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.initial_year. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.data_unit. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.agent_names. Check Documenter's build log for details.

Read entire file

Missing docstring.

Missing docstring for PSRDatabase.file_to_array. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.file_to_array_and_header. Check Documenter's build log for details.

Read registry

Missing docstring.

Missing docstring for PSRDatabase.current_stage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.current_scenario. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.current_block. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.goto. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.next_registry. Check Documenter's build log for details.

File conversion

Missing docstring.

Missing docstring for PSRDatabase.convert_file. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.add_reader!. Check Documenter's build log for details.

Reader mapper

Missing docstring.

Missing docstring for PSRDatabase.ReaderMapper. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.add_reader!. Check Documenter's build log for details.

Modification API

Missing docstring.

Missing docstring for PSRDatabase.create_study. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.create_element!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.set_parm!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.set_vector!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.set_series!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.write_data. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.set_related!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PSRDatabase.set_vector_related!. Check Documenter's build log for details.